+++ src/utils/utils.js
@@ -34,3 +34,14 @@ export function hashFromArray(dataArray) {
   });
   return hash;
 }
+
+export const debounce = (fn, time) => {
+  let timeout;
+
+  return function() {
+    const functionCall = () => fn.apply(this, arguments);
+
+    clearTimeout(timeout);
+    timeout = setTimeout(functionCall, time);
+  };
+};